requirements.txt
numpy==1.23.5
Pillow==9.3.0
tensorflow==2.10.0
tensorflow-hub==0.13.0
fuzzywuzzy==0.18.0
flask==2.2.2
